+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changes to be released in next version
  * MXCrypto: Decrypt events asynchronously and no more on the main thread )(vector-im/element-ios/issues/4306).
  * MXSession: Add the decryptEvents method to decypt a bunch of events asynchronously.
  * MXSession: Make the eventWithEventId method decrypt the event if needed.
+ * MXEventTimeline: Add NSCopying implementation so that another pagination can be done on the same set of data.
 
 🐛 Bugfix
  * MXRoomSummary: Fix decryption of the last message when it is edited (vector-im/element-ios/issues/4322).